Add Auto DevOps compatible Dockerfile
Also nginx.conf to ensure it runs on port 5000 necessary for Auto
DevOps. This should be compatible with Auto DevOps when we transition to
deploying via Auto DevOps.

This could also be useful for running it locally if you do not want to
install all the dependencies. In practice, though, this Dockerfile is
mainly suited for deployments as it is not mounting the directories to
enable ease of local development.