WIP: DO NOT MERGE -- testing old branch name for review app
The purpose of this MR is to see if review app deployment will fail for
the branch name `1-structure-folders-in-storybook`, which is one of the
oldest branch names in the project (https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/merge_requests/2)

**Do not merge**. I will close the MR once the results are in.

Background: Review app deployments are failing unusually frequently, see

    https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/issues/409

One hypothesis is that the cause is the project getting moved causing
MRs opened prior to the move to start failing. But then a very recent MR
failed with seemingly the same symptoms. This would be enough to
reject the hypothesis, except the branch name is `patch-1`, which is
GitLab's default Web-IDE branch name.