docs: changed More information link text to Learn more
The current guidance said to use the phrase More information
to link to docs. Because this wasn't a complete sentence, the
guidance said to use it without a period. To simplify things,
this MR suggests the complete sentence of Learn more.

This MR also recommends using a question to help guide
the user to the docs, while using Learn more as a backup.