feature(documentation): add ability to embed GitLab UI stories
This adds support for including GitLab UI stories in docs pages.

Single stories can be included with the `[[Story:story-id]]` shortcode.
A custom title can be provided by appending `|My title` to the
story ID: `[[Story:story-id|My title]]`.

To provide a Storybook DocsPage to a component's "Implementation" tab,
a `stories` key can be added to the frontmatter:

```
stories:
  - story-id
```