docs(templates): Clean references
As a follow-up to https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/2726, where we renamed the MR template name to `Documentation`. After that change, I noticed that we were referencing it in different places, and they needed to be updated. As part of that, I decided to do a general clean-up of all references to issue and MR templates.