feat: Move to proper static side generation
Before our move to examples via iFrames a lot of markdown and GitLab UI
components needed to be rendered by the browser.

Now that all examples are iFrames, we turn on _proper_ static rendering
of all of our pages. This has the really big advantage that all of the
contents the user see have been precompiled to HTML. This will help with
SEO, but also make the pages load faster.