fix: Add GitLab Pages Auth site to CSP
Sometimes (when logged in?) or during pages deployments you will be
redirected to projects.gitlab.io in order for some Pages Auth checks.
Due to the CSP that process doesn't work out alright when embedded in
an iframe. The fix is simple by adding projects.gitlab.io to our CSP
policy.