refactor: migrate color palettes to grid layouts
This migrates the color palettes away from rows and cols in favor of
grid layouts. The primary goal here is to rely on GitLab UI utils
instead of custom ones.

As part of these changes, the spacing between palettes is being slightly
revisited and the data visualization layout is aligned with the
foundations one.