ci: Run jobs on Merge Request pipelines as well
We start executing the following jobs on merge request pipelines as
well: build, test, code_quality and browser_performance. This can be
done with a `when: always` which overrides the predefined rules. All of
those jobs are supposed to run on the default branch and MRs.

Additionally we can now import the MR workflow of untamper-my-lockfile
and make Danger run only on MRs.