ci: Move review job from Branch to MR pipelines
This one was a little trickier. The review jobs use the branch name
`CI_COMMIT_REF_NAME` in branch pipelines for defining the branch URL.
This variable is not set in MR pipelines, so we need to assign some
value to that variable. Luckily `CI_MERGE_REQUEST_SOURCE_BRANCH_NAME`
should have the same meaning with regards to the review pipelines.