fix: Make anchor links work
When we switched to review apps for pages, we added the `base` property
to the nuxt config. This config seems to be used for a lot of things,
for example to adjust the paths to all assets. But it also makes nuxt
add a `<base>` tag in the header which messes with the page semantics
and header links. We can simply remove it during compilation.

Fixes https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/issues/1754