fix: Untangle the mess which is Error Pages
Nuxt has a mechanism to render a fallback page, which we rendered as
error.html. For some reason which I am a little too lazy to dig up, we
are also creating a separate 404.html for GitLab Pages / NGINX. I guess
because we didn't see proper 404 pages in production?

Anyhow, that being said, let's consolidate the two pages and improve the
logic a little bit. Locally we now will always get the error message
with complete stack traces, in production missing pages will render the
nice 404 site.