ci: Migrate test script to utilize lefthook
We don't need a custom `test.sh` anymore, but what if we used a custom
lefthook job instead?

1. We have lefthook installed anyhow because of our dependencies
2. lefthook is able to run scripts in parallel and give nice output in
   case one or more fail

This still has the benefit that we only need _one_ CI job (because the
overhead of installing dependencies, starting a job, etc.) is much
larger than running these few linting/test scripts.