chore(deps): remove chokidar resolution
This removes the resolution field for `chokidar`. We originally added it
to force `chokidar` to a newer version, `v3.4.0` at the time. We have
been regularly updating the package ever since through Renovate MRs.
However, the recent `v4.0.0` release seems to have broken HMR for
content files. The simplest fix would be to downgrade to the latest `v3`
release, which would be `v3.6.0`. It turns out that removing the
resolution does result in `v3.6.0` being installed, so that should do
it.