fix: Search by hard-coding Pages URL in CI
Before we made the switch to GitLab Pages for our deployment, we used
Pages only for review apps and relied on the CI_PAGES_URL env variable.
This is problematic though, because it is not `design.gitlab.com`, but
actually: https://design-gitlab-com-gitlab-org-<long>.gitlab.io

Now we use this domain for example in order to pull our index for the
search which now will fail due to CORS issues.

If we hard-code the proper URL in CI, we should circumvent the issue.